--- a/Makefile.Common
+++ b/Makefile.Common
@@ -3,7 +3,7 @@ ALL_PKGS := $(sort $(shell go list ./...))
 # COVER_PKGS is the list of packages to include in the coverage
 COVER_PKGS := $(shell go list ./... | tr "\n" ",")
 
-GOTEST_OPT?= -race -timeout 120s -v
+GOTEST_OPT?= -race -timeout 120s
 GOCMD?= go
 GOTEST=$(GOCMD) test
 GOOS := $(shell $(GOCMD) env GOOS)