--- a/exporter/exporterhelper/internal/bounded_memory_queue.go
+++ b/exporter/exporterhelper/internal/bounded_memory_queue.go
@@ -7,7 +7,6 @@ package internal // import "go.opentelemetry.io/collector/exporter/exporterhelpe
 
 import (
 	"context"
-	"runtime/debug"
 	"sync"
 	"sync/atomic"
 
@@ -129,7 +128,6 @@ func (q *boundedMemoryQueue) Produce(item Request) bool {
 	if q.stopped.Load() {
 		return false
 	}
-	debug.PrintStack()
 	waitForAccept := make(chan bool, 1)
 	pipelineItem := &queueEvent{
 		request:    item,