**Description:** 
Fix bounded memory queue stopping behavior to avoid panics.

Before:
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/exporter/exporterhelper/internal
Benchmark_QueueUsage_10000_1_50000-10       	       1	12230623208 ns/op	 2563240 B/op	  149758 allocs/op
Benchmark_QueueUsage_10000_2_50000-10       	       1	6098770209 ns/op	 2563792 B/op	  149763 allocs/op
Benchmark_QueueUsage_10000_5_50000-10       	       1	2288959958 ns/op	 2565400 B/op	  149776 allocs/op
Benchmark_QueueUsage_10000_10_50000-10      	       1	1153033709 ns/op	 2567848 B/op	  149788 allocs/op
Benchmark_QueueUsage_50000_1_50000-10       	       1	59734605167 ns/op	 3201576 B/op	  149753 allocs/op
Benchmark_QueueUsage_50000_2_50000-10       	       1	34550001750 ns/op	 3202728 B/op	  149758 allocs/op
Benchmark_QueueUsage_50000_5_50000-10       	       1	13308404875 ns/op	 3203592 B/op	  149764 allocs/op
Benchmark_QueueUsage_50000_10_50000-10      	       1	7089828500 ns/op	 3205672 B/op	  149783 allocs/op
Benchmark_QueueUsage_10000_1_250000-10      	       1	13404509625 ns/op	12173272 B/op	  749781 allocs/op
Benchmark_QueueUsage_10000_2_250000-10      	       1	6800464208 ns/op	12171432 B/op	  749784 allocs/op
Benchmark_QueueUsage_10000_5_250000-10      	       1	2668322250 ns/op	12169176 B/op	  749779 allocs/op
Benchmark_QueueUsage_10000_10_250000-10     	       1	1297472375 ns/op	12173472 B/op	  749807 allocs/op
BenchmarkPersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_1-10         	  885627	      1408 ns/op	     876 B/op	      11 allocs/op
BenchmarkPersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_10-10        	  386420	      3131 ns/op	    3507 B/op	      11 allocs/op
BenchmarkPersistentStorage_TraceSpans/#traces:_10_#spansPerTrace:_10-10       	   58326	     20666 ns/op	   27779 B/op	      11 allocs/op
PASS
ok  	go.opentelemetry.io/collector/exporter/exporterhelper/internal	166.125s

```


After:
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/exporter/exporterhelper/internal
Benchmark_QueueUsage_10000_1_50000-10       	       1	11422502458 ns/op	 2563368 B/op	  149763 allocs/op
Benchmark_QueueUsage_10000_2_50000-10       	       1	5704231500 ns/op	 2563480 B/op	  149764 allocs/op
Benchmark_QueueUsage_10000_5_50000-10       	       1	2320336334 ns/op	 2566392 B/op	  149785 allocs/op
Benchmark_QueueUsage_10000_10_50000-10      	       1	1141204375 ns/op	 2568200 B/op	  149794 allocs/op
Benchmark_QueueUsage_50000_1_50000-10       	       1	57261643291 ns/op	 3201800 B/op	  149758 allocs/op
Benchmark_QueueUsage_50000_2_50000-10       	       1	28473611125 ns/op	 3201728 B/op	  149760 allocs/op
Benchmark_QueueUsage_50000_5_50000-10       	       1	11412506500 ns/op	 3202040 B/op	  149764 allocs/op
Benchmark_QueueUsage_50000_10_50000-10      	       1	5698795917 ns/op	 3204616 B/op	  149784 allocs/op
Benchmark_QueueUsage_10000_1_250000-10      	       1	11471291042 ns/op	12173264 B/op	  749784 allocs/op
Benchmark_QueueUsage_10000_2_250000-10      	       1	5734358125 ns/op	12169504 B/op	  749789 allocs/op
Benchmark_QueueUsage_10000_5_250000-10      	       1	2311201583 ns/op	12169800 B/op	  749788 allocs/op
Benchmark_QueueUsage_10000_10_250000-10     	       1	1164992208 ns/op	12172664 B/op	  749813 allocs/op
BenchmarkPersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_1-10         	  984412	      1402 ns/op	     854 B/op	      11 allocs/op
BenchmarkPersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_10-10        	  388896	      3226 ns/op	    3506 B/op	      11 allocs/op
BenchmarkPersistentStorage_TraceSpans/#traces:_10_#spansPerTrace:_10-10       	   57034	     21307 ns/op	   27784 B/op	      11 allocs/op
PASS
ok  	go.opentelemetry.io/collector/exporter/exporterhelper/internal	149.059s

```

Benchstat:
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/exporter/exporterhelper/internal
                                                               │  before.txt  │              after6.txt              │
                                                               │    sec/op    │    sec/op     vs base                │
_QueueUsage_10000_1_50000-10                                      12.23 ± ∞ ¹    11.42 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_2_50000-10                                      6.099 ± ∞ ¹    5.704 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_5_50000-10                                      2.289 ± ∞ ¹    2.320 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_10_50000-10                                     1.153 ± ∞ ¹    1.141 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_1_50000-10                                      59.73 ± ∞ ¹    57.26 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_2_50000-10                                      34.55 ± ∞ ¹    28.47 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_5_50000-10                                      13.31 ± ∞ ¹    11.41 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_10_50000-10                                     7.090 ± ∞ ¹    5.699 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_1_250000-10                                     13.40 ± ∞ ¹    11.47 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_2_250000-10                                     6.800 ± ∞ ¹    5.734 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_5_250000-10                                     2.668 ± ∞ ¹    2.311 ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_10_250000-10                                    1.297 ± ∞ ¹    1.165 ± ∞ ¹       ~ (p=1.000 n=1) ²
PersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_1-10     1.408µ ± ∞ ¹   1.402µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_10-10    3.131µ ± ∞ ¹   3.226µ ± ∞ ¹       ~ (p=1.000 n=1) ²
PersistentStorage_TraceSpans/#traces:_10_#spansPerTrace:_10-10   20.67µ ± ∞ ¹   21.31µ ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                                          400.7m         368.4m        -8.06%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                                               │  before.txt   │              after6.txt               │
                                                               │     B/op      │     B/op       vs base                │
_QueueUsage_10000_1_50000-10                                     2.444Mi ± ∞ ¹   2.445Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_2_50000-10                                     2.445Mi ± ∞ ¹   2.445Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_5_50000-10                                     2.447Mi ± ∞ ¹   2.448Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_10_50000-10                                    2.449Mi ± ∞ ¹   2.449Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_1_50000-10                                     3.053Mi ± ∞ ¹   3.053Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_2_50000-10                                     3.054Mi ± ∞ ¹   3.053Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_5_50000-10                                     3.055Mi ± ∞ ¹   3.054Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_10_50000-10                                    3.057Mi ± ∞ ¹   3.056Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_1_250000-10                                    11.61Mi ± ∞ ¹   11.61Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_2_250000-10                                    11.61Mi ± ∞ ¹   11.61Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_5_250000-10                                    11.61Mi ± ∞ ¹   11.61Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_10_250000-10                                   11.61Mi ± ∞ ¹   11.61Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
PersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_1-10       876.0 ± ∞ ¹     854.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
PersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_10-10    3.425Ki ± ∞ ¹   3.424Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
PersistentStorage_TraceSpans/#traces:_10_#spansPerTrace:_10-10   27.13Ki ± ∞ ¹   27.13Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                                          1.100Mi         1.098Mi        -0.18%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                                               │  before.txt  │              after6.txt              │
                                                               │  allocs/op   │  allocs/op    vs base                │
_QueueUsage_10000_1_50000-10                                     149.8k ± ∞ ¹   149.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_2_50000-10                                     149.8k ± ∞ ¹   149.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_5_50000-10                                     149.8k ± ∞ ¹   149.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_10_50000-10                                    149.8k ± ∞ ¹   149.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_1_50000-10                                     149.8k ± ∞ ¹   149.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_2_50000-10                                     149.8k ± ∞ ¹   149.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_50000_5_50000-10                                     149.8k ± ∞ ¹   149.8k ± ∞ ¹       ~ (p=1.000 n=1) ³
_QueueUsage_50000_10_50000-10                                    149.8k ± ∞ ¹   149.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_1_250000-10                                    749.8k ± ∞ ¹   749.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_2_250000-10                                    749.8k ± ∞ ¹   749.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_5_250000-10                                    749.8k ± ∞ ¹   749.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
_QueueUsage_10000_10_250000-10                                   749.8k ± ∞ ¹   749.8k ± ∞ ¹       ~ (p=1.000 n=1) ²
PersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_1-10      11.00 ± ∞ ¹    11.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
PersistentStorage_TraceSpans/#traces:_1_#spansPerTrace:_10-10     11.00 ± ∞ ¹    11.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
PersistentStorage_TraceSpans/#traces:_10_#spansPerTrace:_10-10    11.00 ± ∞ ¹    11.00 ± ∞ ¹       ~ (p=1.000 n=1) ³
geomean                                                          34.29k         34.29k        +0.00%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal
```
**Link to tracking Issue:** 
Fixes #8124