@atoulme 2023-11-12
Closing as not the right approach. I discussed this at KubeCon with Bogdan and my approach is wrong here. There are a couple follow-ups we can do on top of this to make the code more maintainable:
* Do not let this queue be created without consumers or no capacity (some tests used to do that, and that really confused me, even though in practice you're not allowed). I can look at Validate() to make sure I use the same checks.
* Make the benchmarks associated with this effort run with TB, and do not use a time.Sleep in the callback. We can block on the callback instead or use a wait group to make sure all items are consumed properly.